simpledae refactor
