+++ src/ibex_bluesky_core/devices/dae/__init__.py
@@ -30,7 +30,7 @@ from ibex_bluesky_core.devices.dae._period_settings import (
     SinglePeriodSettings,
 )
 from ibex_bluesky_core.devices.dae._settings import DaeSettings, DaeSettingsData, DaeTimingSource
-from ibex_bluesky_core.devices.dae._spectra import DaeSpectra
+from ibex_bluesky_core.devices.dae._spectra import DaeSpectra, PolarisedWavelengthBand, WavelengthBand
 from ibex_bluesky_core.devices.dae._tcb_settings import (
     DaeTCBSettings,
     DaeTCBSettingsData,
@@ -67,6 +67,8 @@ __all__ = [
     "TimeRegime",
     "TimeRegimeMode",
     "TimeRegimeRow",
+    "PolarisedWavelengthBand",
+    "WavelengthBand"
 ]
 
 T = TypeVar("T", bound=SignalDatatype)
