+++ src/ibex_bluesky_core/devices/simpledae/_waiters.py
@@ -11,19 +11,12 @@ from ophyd_async.core import (
     wait_for_value,
 )
 
+from ibex_bluesky_core.devices.simpledae._strategies import Waiter
 from ibex_bluesky_core.devices.dae import Dae
-from ibex_bluesky_core.devices.dae.strategies._extra_readables import ProvidesExtraReadables
 
 logger = logging.getLogger(__name__)
 
 T = TypeVar("T", int, float)
-
-
-class Waiter(ProvidesExtraReadables):
-    """Waiter specifies how the dae will wait for a scan point to complete counting."""
-
-    async def wait(self, dae: Dae) -> None:
-        """Wait for the acquisition to complete."""
 
 
 class SimpleWaiter(Waiter, Generic[T], ABC):
