+++ tests/devices/strategies/test_reducers.py
@@ -8,7 +8,7 @@ import pytest
 import scipp as sc
 from ophyd_async.testing import get_mock_put, set_mock_value
 
-from ibex_bluesky_core.devices.dae.strategies import (
+from ibex_bluesky_core.devices.simpledae import (
     VARIANCE_ADDITION,
     GoodFramesNormalizer,
     MonitorNormalizer,
