+++ tests/devices/strategies/test_waiters.py
@@ -3,7 +3,7 @@ import asyncio
 import pytest
 from ophyd_async.testing import set_mock_value
 
-from ibex_bluesky_core.devices.dae.strategies import (
+from ibex_bluesky_core.devices.simpledae import (
     GoodFramesWaiter,
     GoodUahWaiter,
     MEventsWaiter,
