+++ tests/plans/test_det_map_align.py
@@ -6,7 +6,7 @@ import pytest
 from ophyd_async.core import soft_signal_rw
 from ophyd_async.testing import callback_on_mock_put, set_mock_value
 
-from ibex_bluesky_core.devices.dae.strategies import (
+from ibex_bluesky_core.devices.simpledae import (
     PeriodPerPointController,
     PeriodSpecIntegralsReducer,
     Waiter,
